5d210a2 [REFACTOR] ./flow/* : Refactoring

+++ src/main/java/com/pos/app/flow/OrderFlow.java
@@ -10,18 +10,18 @@ public class OrderFlow {
     @Autowired
     ProductService productService;
 
-    public int getTotal_price(int barcode, int quantity){
+    public Double getTotal_price(Integer barcode, Integer quantity){
         ProductPojo productPojo = productService.getProductByBarcode(barcode);
         return productPojo.getProductPrice()*quantity;
     }
-    public int getClient_id(int barcode){
+    public Integer getClient_id(Integer barcode){
         ProductPojo productPojo = productService.getProductByBarcode(barcode);
         return productPojo.getClientId();
     }
-    public String getProductName(int barcode) {
+    public String getProductName(Integer barcode) {
         return productService.getProductByBarcode(barcode).getProductName();
     }
-    public void reduceInventory(int barcode, int quantity){
+    public void reduceInventory(Integer barcode, Integer quantity){
         ProductPojo productPojo = productService.getProductByBarcode(barcode);
         productPojo.setProductQuantity(productPojo.getProductQuantity() - quantity);
         productService.updateProduct(productPojo.getProductId(),productPojo);
